Created a component for status logs.
